[Eye Colors] Visual Bug Fix
Need to swap material slots manually in blender. It is possible to proceed by looking only to change the pupil slots in script but to handle information consistently, changing the models to be consistent is the best way to handle this, making sure new eye models follow the same structure.